Improvements to large imports, currency changes, and initial, clean state.
If you have no expenses/budgets, the app won't appear to be in a broken loading state anymore. If you import thousands and have a Sync Token, the app will do the import in chunks, avoiding potential rate limits.

When you change the currency, it'll also be immediately visible, and will no longer require you to close the app and open it again.